Move shared logic to highlight component
