+++ src/documentation/ToolkitDocumentation/Highlight.tsx
@@ -1,19 +1,81 @@
-import { Box, BoxProps } from "@chakra-ui/react";
+import { Box, BoxProps } from "@chakra-ui/layout";
+import { usePrefersReducedMotion, usePrevious } from "@chakra-ui/react";
+import { useEffect, useRef, useState, useCallback } from "react";
+import { Anchor } from "../../router-hooks";
+import { useLogging } from "../../logging/logging-hooks";
+import { useScrollablePanelAncestor } from "../../workbench/ScrollablePanel";
 
 interface HighlightProps extends BoxProps {
-  value: boolean;
+  anchor?: Anchor;
+  entryName: string;
+  active: Boolean | undefined;
+  disclosure: any;
 }
 
-const Highlight = ({ value, children, ...props }: HighlightProps) => {
-  const style = value
+const Highlight = ({
+  children,
+  active,
+  anchor,
+  entryName,
+  disclosure,
+  ...props
+}: HighlightProps) => {
+  const ref = useRef<HTMLDivElement>(null);
+  const previousAnchor = usePrevious(anchor);
+  const scrollable = useScrollablePanelAncestor();
+  const prefersReducedMotion = usePrefersReducedMotion();
+  const logging = useLogging();
+  const [highlighting, setHighlighting] = useState(false);
+  useEffect(() => {
+    if (previousAnchor !== anchor && active) {
+      logging.log("Activating " + entryName);
+      disclosure.onOpen();
+      // Delay until after the opening animation so the full container height is known for the scroll.
+      window.setTimeout(() => {
+        if (ref.current && scrollable.current) {
+          scrollable.current.scrollTo({
+            // Fudge to account for the fixed header and to leave a small gap.
+            top: ref.current.offsetTop - 112 - 25,
+            behavior: prefersReducedMotion ? "auto" : "smooth",
+          });
+        }
+        setTimeout(() => {
+          setHighlighting(true);
+          setTimeout(() => {
+            setHighlighting(false);
+          }, 3000);
+        }, 300);
+      }, 150);
+    }
+  }, [
+    anchor,
+    scrollable,
+    active,
+    previousAnchor,
+    prefersReducedMotion,
+    logging,
+    disclosure,
+    entryName,
+  ]);
+
+  const handleHighlightClick = useCallback(() => {
+    setHighlighting(false);
+  }, [setHighlighting]);
+
+  const style = highlighting
     ? {
         backgroundColor: "brand.100",
         transition: "background-color ease-out 0.2s",
       }
     : { transition: "background-color ease-in 0.6s" };
   return (
-    <Box {...props} {...style} borderLeftRadius="md">
-      {children}
+    <Box
+      onClick={handleHighlightClick}
+      borderLeftRadius="md"
+      {...props}
+      {...style}
+    >
+      <Box ref={ref}>{children}</Box>
     </Box>
   );
 };
